// confmod-app/frontend/src/app/openapi/model/rawObservationSpecInput.ts
/**
 * Confmod Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RawFeatureSpec } from './rawFeatureSpec';


export interface RawObservationSpecInput { 
    type: RawObservationSpecInput.TypeEnum;
    description?: string;
    features: Array<RawFeatureSpec>;
}
export namespace RawObservationSpecInput {
    export type TypeEnum = 'E' | 'M';
    export const TypeEnum = {
        E: 'E' as TypeEnum,
        M: 'M' as TypeEnum
    };
}